Move OpenDrive related classes to its own folder
